Bug 1590762 - Part 2: Bump the id for channel registration to uint64_t. r=mattwoodrow,necko-reviewers,valentin
This patch also makes the identifier for channels global, in the sense
that the generated identifier is generated outside of and passed to
the nsIRedirectChannelRegistrar.

Differential Revision: https://phabricator.services.mozilla.com/D79820

UltraBlame original commit: b1506f4ded8177ff08c39e82dbf2a003360c529a

+++ docshell/base/BrowsingContext.cpp
@@ -1585,7 +1585,7 @@ nsresult BrowsingContext::LoadURI(nsDocShellLoadState* aLoadState,
     if (ContentParent* cp = Canonical()->GetContentParent()) {
       
       
-      uint32_t loadIdentifier = 0;
+      uint64_t loadIdentifier = 0;
       if (Canonical()->AttemptLoadURIInParent(aLoadState, &loadIdentifier)) {
         aLoadState->SetLoadIdentifier(loadIdentifier);
       }
